+++ test/testPublisher.c
@@ -6,7 +6,6 @@
 #include <libgen.h>
 #include <zconf.h>
 #include <stdlib.h>
-#include <time.h>
 
 #include "csw/csw.h"
 
@@ -21,19 +20,29 @@
  *  sbt stage
  *  test-deploy/target/universal/stage/bin/test-container-cmd-app --local test-deploy/src/main/resources/TestContainer.conf
  *
- * The test assembly subscribes to the events published by this test.
+ * Note: The test assembly subscribes to exactly the named events published by this test,
+ * so if you change any event names, update the assembly code in TestAssemblyHandlers.
  */
 
 static char *prefix = "CSW.TestPublisher";
 
 // Tests publishing a simple event with multiple values
 static void publishSimpleEvent(CswEventServiceContext publisher) {
-    double ar[] = {1.1, 2.2, 3.3};
-    CswArrayValue values = {.values = ar, .numValues = 3};
-    CswParameter param = cswMakeParameter("DoubleValue", DoubleKey, values, "arcmin");
-    CswParameter params[] = {param};
-    CswParamSet paramSet = {.params = params, .numParams = 1};
-    CswEvent event = cswMakeEvent(SystemEvent, prefix, "SimpleDoubleEvent", paramSet);
+    double ar1[] = {1.1, 2.2, 3.3};
+    CswArrayValue values1 = {.values = ar1, .numValues = 3};
+    CswParameter param1 = cswMakeParameter("DoubleValue", DoubleKey, values1, csw_unit_arcmin);
+
+    char* ar2 = {"one"};
+    CswArrayValue values2 = {.values = ar2, .numValues = 1};
+    CswParameter param2 = cswMakeParameter("StringValue", StringKey, values2, csw_unit_NoUnits);
+
+    char* ar3 = {"choice1"};
+    CswArrayValue values3 = {.values = ar3, .numValues = 1};
+    CswParameter param3 = cswMakeParameter("ChoiceValue", ChoiceKey, values3, csw_unit_NoUnits);
+
+    CswParameter params[] = {param1, param2, param3};
+    CswParamSet paramSet = {.params = params, .numParams = 3};
+    CswEvent event = cswMakeEvent(SystemEvent, prefix, "SimpleEvent", paramSet);
     cswEventPublish(publisher, event);
     cswFreeEvent(event);
 }
@@ -44,7 +53,7 @@ static void publishUtcTimeEvent(CswEventServiceContext publisher) {
     // Use fixed date so that test comparison works
     CswUtcTime ar[] = {cswMakeUtcTime(1625066893, 372333847)};
     CswArrayValue values = {.values = ar, .numValues = 1};
-    CswParameter param = cswMakeParameter("utcTimeValue", UTCTimeKey, values, "NoUnits");
+    CswParameter param = cswMakeParameter("utcTimeValue", UTCTimeKey, values, csw_unit_NoUnits);
     CswParameter params[] = {param};
     CswParamSet paramSet = {.params = params, .numParams = 1};
     CswEvent event = cswMakeEvent(SystemEvent, prefix, "UtcTimeEvent", paramSet);
@@ -58,7 +67,7 @@ static void publishTaiTimeEvent(CswEventServiceContext publisher) {
     // Use fixed date so that test comparison works
     CswTaiTime ar[] = {cswMakeTaiTime(1625066980, 763689367)};
     CswArrayValue values = {.values = ar, .numValues = 1};
-    CswParameter param = cswMakeParameter("taiTimeValue", TAITimeKey, values, "NoUnits");
+    CswParameter param = cswMakeParameter("taiTimeValue", TAITimeKey, values, csw_unit_NoUnits);
     CswParameter params[] = {param};
     CswParamSet paramSet = {.params = params, .numParams = 1};
     CswEvent event = cswMakeEvent(SystemEvent, prefix, "TaiTimeEvent", paramSet);
@@ -72,7 +81,7 @@ static void publishInts(CswEventServiceContext publisher) {
     // -- IntKey parameter contains one or more int values --
     int intValues[] = {42, 43};
     CswArrayValue arrayValues1 = {.values = intValues, .numValues = 2};
-    CswParameter intParam = cswMakeParameter("IntValue", IntKey, arrayValues1, "arcsec");
+    CswParameter intParam = cswMakeParameter("IntValue", IntKey, arrayValues1, csw_unit_arcsec);
 
     // -- IntArrayKey parameter contains one or more int array values --
     int intArrayValues[2][4] = {
@@ -84,7 +93,7 @@ static void publishInts(CswEventServiceContext publisher) {
             "IntArrayValue",
             IntArrayKey,
             makeArrayValues((void **) intArrayValues, 2, intArrayHolder, 4),
-            "arcsec");
+            csw_unit_arcsec);
 
 
     // -- IntMatrixKey parameter contains one or more int matrix values (Sorry, this gets ugly in C...) --
@@ -108,7 +117,7 @@ static void publishInts(CswEventServiceContext publisher) {
     };
     CswArrayValue matrixValues = {.arrayValues = ar, .numValues = 2};
 
-    CswParameter intMatrixParam = cswMakeParameter("IntMatrixValue", IntMatrixKey, matrixValues, "arcsec");
+    CswParameter intMatrixParam = cswMakeParameter("IntMatrixValue", IntMatrixKey, matrixValues, csw_unit_arcsec);
 
     // -- ParamSet
     CswParameter params[] = {intParam, intArrayParam, intMatrixParam};
@@ -128,7 +137,7 @@ static void publishDoubles(CswEventServiceContext publisher) {
     // -- DoubleKey parameter contains one or more double values --
     double doubleValues[] = {42.1, 43.5};
     CswArrayValue arrayValues1 = {.values = doubleValues, .numValues = 2};
-    CswParameter doubleParam = cswMakeParameter("DoubleValue", DoubleKey, arrayValues1, "arcsec");
+    CswParameter doubleParam = cswMakeParameter("DoubleValue", DoubleKey, arrayValues1, csw_unit_arcsec);
 
     // -- DoubleArrayKey parameter contains one or more double array values --
     double doubleArrayValues[2][4] = {
@@ -140,7 +149,7 @@ static void publishDoubles(CswEventServiceContext publisher) {
             "DoubleArrayValue",
             DoubleArrayKey,
             makeArrayValues((void **) doubleArrayValues, 2, doubleArrayHolder, 4),
-            "arcsec");
+            csw_unit_arcsec);
 
 
     // -- DoubleMatrixKey parameter contains one or more double matrix values (Sorry, this gets ugly in C...) --
@@ -164,7 +173,7 @@ static void publishDoubles(CswEventServiceContext publisher) {
     };
     CswArrayValue matrixValues = {.arrayValues = ar, .numValues = 2};
 
-    CswParameter doubleMatrixParam = cswMakeParameter("DoubleMatrixValue", DoubleMatrixKey, matrixValues, "arcsec");
+    CswParameter doubleMatrixParam = cswMakeParameter("DoubleMatrixValue", DoubleMatrixKey, matrixValues, csw_unit_arcsec);
 
     // -- ParamSet
     CswParameter params[] = {doubleParam, doubleArrayParam, doubleMatrixParam};
@@ -187,7 +196,7 @@ static void publishAltAzCoord(CswEventServiceContext publisher) {
     CswCoord values[1];
     values[0].altAzCoord = altAzCoord1;
     CswArrayValue arrayValues = {.values = values, .numValues = 1};
-    CswParameter coordParam = cswMakeParameter("CoordParam", CoordKey, arrayValues, "NoUnits");
+    CswParameter coordParam = cswMakeParameter("CoordParam", CoordKey, arrayValues, csw_unit_NoUnits);
 
     // -- ParamSet
     CswParameter params[] = {coordParam};
@@ -208,7 +217,7 @@ static void publishCometCoord(CswEventServiceContext publisher) {
     CswCoord values[1];
     values[0].cometCoord = cometCoord1;
     CswArrayValue arrayValues = {.values = values, .numValues = 1};
-    CswParameter coordParam = cswMakeParameter("CoordParam", CoordKey, arrayValues, "NoUnits");
+    CswParameter coordParam = cswMakeParameter("CoordParam", CoordKey, arrayValues, csw_unit_NoUnits);
 
     // -- ParamSet
     CswParameter params[] = {coordParam};
@@ -230,7 +239,7 @@ static void publishMinorPlanetCoord(CswEventServiceContext publisher) {
     CswCoord values[1];
     values[0].minorPlanetCoord = minorPlanetCoord1;
     CswArrayValue arrayValues = {.values = values, .numValues = 1};
-    CswParameter coordParam = cswMakeParameter("CoordParam", CoordKey, arrayValues, "NoUnits");
+    CswParameter coordParam = cswMakeParameter("CoordParam", CoordKey, arrayValues, csw_unit_NoUnits);
 
     // -- ParamSet
     CswParameter params[] = {coordParam};
@@ -253,7 +262,7 @@ static void publishSolarSystemCoords(CswEventServiceContext publisher) {
     values[0].solarSystemCoord = solarSystemCoord1;
     values[1].solarSystemCoord = solarSystemCoord2;
     CswArrayValue arrayValues = {.values = values, .numValues = 2};
-    CswParameter coordParam = cswMakeParameter("CoordParam", CoordKey, arrayValues, "NoUnits");
+    CswParameter coordParam = cswMakeParameter("CoordParam", CoordKey, arrayValues, csw_unit_NoUnits);
 
     // -- ParamSet
     CswParameter params[] = {coordParam};
@@ -276,7 +285,7 @@ static void publishEqCoords(CswEventServiceContext publisher) {
     values[0].eqCoord = eqCoord1;
     values[1].eqCoord = eqCoord2;
     CswArrayValue arrayValues = {.values = values, .numValues = 2};
-    CswParameter coordParam = cswMakeParameter("CoordParam", CoordKey, arrayValues, "NoUnits");
+    CswParameter coordParam = cswMakeParameter("CoordParam", CoordKey, arrayValues, csw_unit_NoUnits);
 
     // -- ParamSet
     CswParameter params[] = {coordParam};
@@ -296,19 +305,19 @@ static void publishStruct(CswEventServiceContext publisher) {
     // Double parameter
     double ar1[] = {1.1, 2.2, 3.3};
     CswArrayValue values1 = {.values = ar1, .numValues = 3};
-    CswParameter param1 = cswMakeParameter("doubleEventValues", DoubleKey, values1, "arcmin");
+    CswParameter param1 = cswMakeParameter("doubleEventValues", DoubleKey, values1, csw_unit_arcmin);
 
     // Int parameter
     int ar2[] = {1, 2, 3};
     CswArrayValue values2 = {.values = ar2, .numValues = 3};
-    CswParameter param2 = cswMakeParameter("intEventValues", IntKey, values2, "arcmin");
+    CswParameter param2 = cswMakeParameter("intEventValues", IntKey, values2, csw_unit_arcmin);
 
     // Struct parameter containing two fields (the two above parameters)
     CswParameter params1[] = {param1, param2};
     CswParamSet paramSet1 = {.params = params1, .numParams = 2};
     CswParamSet ar[] = {paramSet1};
     CswArrayValue values = {.values = ar, .numValues = 1};
-    CswParameter param = cswMakeParameter("structEventValues", StructKey, values, "NoUnits");
+    CswParameter param = cswMakeParameter("structEventValues", StructKey, values, csw_unit_NoUnits);
 
     // Param set for the event
     CswParameter params[] = {param};
@@ -322,6 +331,8 @@ static void publishStruct(CswEventServiceContext publisher) {
 }
 
 
+#pragma clang diagnostic push
+#pragma ide diagnostic ignored "UnusedParameter"
 int main(int argc, char **argv) {
     CswEventServiceContext publisher = cswEventPublisherInit();
 
@@ -356,4 +367,5 @@ int main(int argc, char **argv) {
 
     return status;
 }
+#pragma clang diagnostic pop