+++ test/testSupport/test-assembly/src/main/scala/org/tmt/csw/testassembly/TestAssemblyHandlers.scala
@@ -27,7 +27,7 @@ object TestAssemblyHandlers {
   // Event published by C code and subscribed to here
   private val prefix = Prefix(CSW, "TestPublisher")
   private val eventNames = List(
-    "SimpleDoubleEvent",
+    "SimpleEvent",
     "UtcTimeEvent",
     "TaiTimeEvent",
     "IntArrayMatrixEvent",